+++ R/tidy_add_header_rows.R
@@ -29,35 +29,35 @@
 #' @family tidy_helpers
 #' @examplesIf interactive()
 #' if (.assert_package("gtsummary", boolean = TRUE)) {
-#' df <- Titanic %>%
-#'   dplyr::as_tibble() %>%
-#'   dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
+#'   df <- Titanic %>%
+#'     dplyr::as_tibble() %>%
+#'     dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
 #'
-#' res <- df %>%
+#'   res <- df %>%
+#'     glm(
+#'       Survived ~ Class + Age + Sex,
+#'       data = ., weights = .$n, family = binomial,
+#'       contrasts = list(Age = contr.sum, Class = "contr.SAS")
+#'     ) %>%
+#'     tidy_and_attach() %>%
+#'     tidy_add_variable_labels(labels = list(Class = "Custom label for Class")) %>%
+#'     tidy_add_reference_rows()
+#'   res %>% tidy_add_header_rows()
+#'   res %>% tidy_add_header_rows(show_single_row = all_dichotomous())
+#'
 #'   glm(
-#'     Survived ~ Class + Age + Sex,
-#'     data = ., weights = .$n, family = binomial,
-#'     contrasts = list(Age = contr.sum, Class = "contr.SAS")
+#'     response ~ stage + grade * trt,
+#'     gtsummary::trial,
+#'     family = binomial,
+#'     contrasts = list(
+#'       stage = contr.treatment(4, base = 3),
+#'       grade = contr.treatment(3, base = 2),
+#'       trt = contr.treatment(2, base = 2)
+#'     )
 #'   ) %>%
-#'   tidy_and_attach() %>%
-#'   tidy_add_variable_labels(labels = list(Class = "Custom label for Class")) %>%
-#'   tidy_add_reference_rows()
-#' res %>% tidy_add_header_rows()
-#' res %>% tidy_add_header_rows(show_single_row = all_dichotomous())
-#'
-#' glm(
-#'   response ~ stage + grade * trt,
-#'   gtsummary::trial,
-#'   family = binomial,
-#'   contrasts = list(
-#'     stage = contr.treatment(4, base = 3),
-#'     grade = contr.treatment(3, base = 2),
-#'     trt = contr.treatment(2, base = 2)
-#'   )
-#' ) %>%
-#'   tidy_and_attach() %>%
-#'   tidy_add_reference_rows() %>%
-#'   tidy_add_header_rows()
+#'     tidy_and_attach() %>%
+#'     tidy_add_reference_rows() %>%
+#'     tidy_add_header_rows()
 #' }
 tidy_add_header_rows <- function(x,
                                  show_single_row = NULL,
@@ -72,11 +72,12 @@ tidy_add_header_rows <- function(x,
   }
 
   if ("header_row" %in% names(x)) {
-    if (!quiet)
+    if (!quiet) {
       cli_alert_danger(paste(
         "{.code tidy_add_header_rows()} has already been applied.",
         "x has been returned unchanged."
       ))
+    }
     return(x)
   }
 
@@ -97,36 +98,58 @@ tidy_add_header_rows <- function(x,
   )
 
   has_reference_row <- "reference_row" %in% names(x)
-  if (!has_reference_row)
+  if (!has_reference_row) {
     x$reference_row <- FALSE
+  }
 
   xx <- x
-  if ("y.level" %in% names(x) &&
+  if (
+    "y.level" %in% names(x) &&
       # specific case for multinomial models
-      (inherits(model, "multinom") || inherits(model, "LORgee"))) {
+      (inherits(model, "multinom") || inherits(model, "LORgee"))
+  ) {
     xx <- xx %>%
       dplyr::filter(.data$y.level == x$y.level[1])
   }
 
   # checking if variables incorrectly requested for single row summary
-  bad_single_row <- xx %>%
-    dplyr::filter(!is.na(.data$variable),
-                  is.na(.data$reference_row) | !.data$reference_row,
-                  .data$variable %in% show_single_row) %>%
-    dplyr::group_by(.data$variable) %>%
-    dplyr::count() %>%
-    dplyr::filter(.data$n > 1) %>%
-    dplyr::pull(.data$variable)
+  if ("component" %in% colnames(xx)) {
+    bad_single_row <- xx %>%
+      dplyr::filter(
+        !is.na(.data$variable),
+        is.na(.data$reference_row) | !.data$reference_row,
+        .data$variable %in% show_single_row
+      ) %>%
+      dplyr::group_by(.data$component, .data$variable) %>%
+      dplyr::count() %>%
+      dplyr::filter(.data$n > 1) %>%
+      dplyr::pull(.data$variable)
+  } else {
+    bad_single_row <- xx %>%
+      dplyr::filter(
+        !is.na(.data$variable),
+        is.na(.data$reference_row) | !.data$reference_row,
+        .data$variable %in% show_single_row
+      ) %>%
+      dplyr::group_by(.data$variable) %>%
+      dplyr::count() %>%
+      dplyr::filter(.data$n > 1) %>%
+      dplyr::pull(.data$variable)
+  }
   if (length(bad_single_row) > 0) {
-    if (!quiet)
-      paste("Variable(s) {paste(shQuote(bad_single_row), collapse = \", \")} were",
-            "incorrectly requested to be printed on a single row.") %>%
-      cli_alert_danger()
-    if (strict)
+    if (!quiet) {
+      paste(
+        "Variable(s) {paste(shQuote(bad_single_row), collapse = \", \")} were",
+        "incorrectly requested to be printed on a single row."
+      ) %>%
+        cli_alert_danger()
+    }
+    if (strict) {
       cli::cli_abort(
         "Incorrect call with `show_single_row=`. Quitting execution.",
         call = NULL
       )
+    }
     show_single_row <- setdiff(show_single_row, bad_single_row)
   }
 
@@ -134,14 +157,24 @@ tidy_add_header_rows <- function(x,
     length(show_single_row) > 0 &&
       any(x$variable %in% show_single_row)
   ) {
-
-    variables_to_simplify <- xx %>%
-      dplyr::filter(
-        .data$variable %in% show_single_row & !.data$reference_row
-      ) %>%
-      dplyr::count(.data$variable) %>%
-      dplyr::filter(.data$n == 1) %>%
-      purrr::pluck("variable")
+    if ("component" %in% colnames(xx)) {
+      variables_to_simplify <- xx %>%
+        dplyr::filter(
+          .data$variable %in% show_single_row & !.data$reference_row
+        ) %>%
+        dplyr::count(.data$component, .data$variable) %>%
+        dplyr::filter(.data$n == 1) %>%
+        purrr::pluck("variable") %>%
+        unique()
+    } else {
+      variables_to_simplify <- xx %>%
+        dplyr::filter(
+          .data$variable %in% show_single_row & !.data$reference_row
+        ) %>%
+        dplyr::count(.data$variable) %>%
+        dplyr::filter(.data$n == 1) %>%
+        purrr::pluck("variable")
+    }
 
     # removing reference rows of those variables
     if (length(variables_to_simplify) > 0) {
@@ -165,8 +198,9 @@ tidy_add_header_rows <- function(x,
       )
   }
 
-  if (!has_reference_row)
+  if (!has_reference_row) {
     x <- x %>% dplyr::select(-dplyr::all_of("reference_row"))
+  }
 
   # computing header rows ---------------
 
@@ -175,9 +209,11 @@ tidy_add_header_rows <- function(x,
       rank = seq_len(dplyr::n()) # for sorting table at the end
     )
 
-  if ("y.level" %in% names(x) &&
+  if (
+    "y.level" %in% names(x) &&
       # specific case for multinomial models
-      (inherits(model, "multinom") || inherits(model, "LORgee"))) {
+      (inherits(model, "multinom") || inherits(model, "LORgee"))
+  ) {
     header_rows <- x %>%
       dplyr::filter(!is.na(.data$variable) & !.data$variable %in% show_single_row)
 
@@ -185,6 +221,33 @@ tidy_add_header_rows <- function(x,
       header_rows <- header_rows %>%
         dplyr::mutate(term_cleaned = .clean_backticks(.data$term, .data$variable)) %>%
         dplyr::group_by(.data$variable, .data$y.level) %>%
+        dplyr::summarise(
+          var_class = dplyr::first(.data$var_class),
+          var_type = dplyr::first(.data$var_type),
+          var_label = dplyr::first(.data$var_label),
+          var_nlevels = dplyr::first(.data$var_nlevels),
+          contrasts = dplyr::first(.data$contrasts),
+          contrasts_type = dplyr::first(.data$contrasts_type),
+          var_nrow = dplyr::n(),
+          var_test = sum(.data$term_cleaned != .data$variable),
+          rank = min(.data$rank) - .25,
+          .groups = "drop_last"
+        ) %>%
+        dplyr::filter(.data$var_nrow >= 2 | .data$var_test > 0) %>%
+        dplyr::select(-dplyr::all_of(c("var_nrow", "var_test"))) %>%
+        dplyr::mutate(
+          header_row = TRUE,
+          label = .data$var_label
+        )
+    }
+  } else if ("component" %in% names(x)) {
+    header_rows <- x %>%
+      dplyr::filter(!is.na(.data$variable) & !.data$variable %in% show_single_row)
+
+    if (nrow(header_rows) > 0) {
+      header_rows <- header_rows %>%
+        dplyr::mutate(term_cleaned = .clean_backticks(.data$term, .data$variable)) %>%
+        dplyr::group_by(.data$variable, .data$component) %>%
         dplyr::summarise(
           var_class = dplyr::first(.data$var_class),
           var_type = dplyr::first(.data$var_type),
@@ -212,7 +275,7 @@ tidy_add_header_rows <- function(x,
           !.data$var_type %in% c("ran_pars", "ran_vals")
       )
 
-    if (nrow(header_rows) > 0)
+    if (nrow(header_rows) > 0) {
       header_rows <- header_rows %>%
         dplyr::mutate(term_cleaned = .clean_backticks(.data$term, .data$variable)) %>%
         dplyr::group_by(.data$variable) %>%
@@ -235,6 +298,7 @@ tidy_add_header_rows <- function(x,
           header_row = TRUE,
           label = .data$var_label
         )
+    }
   }
 
   x <- x %>%