+++ R/tidy_disambiguate_terms.R
@@ -17,13 +17,14 @@
 #' @examplesIf interactive()
 #' if (
 #'   .assert_package("lme4", boolean = TRUE) &&
-#'   .assert_package("broom.mixed", boolean = TRUE) &&
-#'   .assert_package("gtsummary", boolean = TRUE)
+#'     .assert_package("broom.mixed", boolean = TRUE) &&
+#'     .assert_package("gtsummary", boolean = TRUE)
 #' ) {
-#' mod <- lme4::lmer(marker ~ stage + (1|grade) + (death|response), gtsummary::trial)
-#' mod %>% tidy_and_attach() %>% tidy_disambiguate_terms()
+#'   mod <- lme4::lmer(marker ~ stage + (1 | grade) + (death | response), gtsummary::trial)
+#'   mod %>%
+#'     tidy_and_attach() %>%
+#'     tidy_disambiguate_terms()
 #' }
-
 tidy_disambiguate_terms <- function(x, sep = ".", model = tidy_get_model(x), quiet = FALSE) {
   if (is.null(model)) {
     cli::cli_abort(c(
@@ -31,10 +32,13 @@ tidy_disambiguate_terms <- function(x, sep = ".", model = tidy_get_model(x), qui
       "You need to pass it or to use {.fn tidy_and_attach}."
     ))
   }
-
   if ("original_term" %in% names(x)) {
-    if (!quiet && !inherits(model, "LORgee")) {
-      # no alert for multgee models
+    if (
+      !quiet &&
+        !inherits(model, "LORgee") && # no alert for multgee models
+        !inherits(model, "zeroinfl") && # or zeroninfl/hurdle
+        !inherits(model, "hurdle")
+    ) {
       cli_alert_danger(paste(
         "{.code tidy_disambiguate_terms()} has already been applied.",
         "x has been returned unchanged."